+++ projects/igniteui-angular/migrations/update-16_1_0/index.spec.ts
@@ -132,4 +132,58 @@ describe(`Update to ${version}`, () => {
         `
         );
     });
+
+    it('Should remove multiSelection property set to "true" and replace it with selectionMode property set to "multi"', async () => {
+        appTree.create(
+            `/testSrc/appPrefix/component/test.component.html`,
+            `<igx-buttongroup [multiSelection]="true">
+                <button igxButton>
+                    Button 1
+                </button>
+                <button igxButton>
+                    Button 2
+                </button>
+            </igx-buttongroup>`);
+
+        const tree = await schematicRunner.runSchematicAsync(migrationName, {}, appTree)
+            .toPromise();
+    
+        expect(tree.readContent('/testSrc/appPrefix/component/test.component.html'))
+            .toEqual(
+            `<igx-buttongroup [selectionMode]="'multi'">
+                <button igxButton>
+                    Button 1
+                </button>
+                <button igxButton>
+                    Button 2
+                </button>
+            </igx-buttongroup>`);
+    });
+
+    it('Should remove multiSelection property set to "false"', async () => {
+        appTree.create(
+            '/testSrc/appPrefix/component/test.component.html',
+            `<igx-buttongroup [multiSelection]="false">
+                <button igxButton>
+                    Button 1
+                </button>
+                <button igxButton>
+                    Button 2
+                </button>
+            </igx-buttongroup>`);
+
+        const tree = await schematicRunner.runSchematicAsync(migrationName, {}, appTree)
+            .toPromise();
+
+        expect(tree.readContent('/testSrc/appPrefix/component/test.component.html'))
+            .toEqual(
+            `<igx-buttongroup >
+                <button igxButton>
+                    Button 1
+                </button>
+                <button igxButton>
+                    Button 2
+                </button>
+            </igx-buttongroup>`);
+    });
 });