+++ projects/igniteui-angular/src/lib/buttonGroup/buttongroup.component.spec.ts
@@ -5,6 +5,7 @@ import { configureTestSuite } from '../test-utils/configure-suite';
 import { NoopAnimationsModule } from '@angular/platform-browser/animations';
 import { UIInteractions } from '../test-utils/ui-interactions.spec';
 import { IgxButtonDirective } from '../directives/button/button.directive';
+import { getComponentSize } from '../core/utils';
 
 interface IButton {
     type?: string;
@@ -70,7 +71,7 @@ describe('IgxButtonGroup', () => {
         expect(instance.buttonGroup.id).toContain('igx-buttongroup-');
         expect(buttongroup.disabled).toBeFalsy();
         expect(buttongroup.alignment).toBe(ButtonGroupAlignment.horizontal);
-        expect(buttongroup.multiSelection).toBeFalsy();
+        expect(buttongroup.selectionMode).toBe('single');
         expect(buttongroup.itemContentCssClass).toBeUndefined();
         expect(buttongroup.selectedIndexes.length).toEqual(1);
         expect(buttongroup.selectedButtons.length).toEqual(1);
@@ -87,7 +88,7 @@ describe('IgxButtonGroup', () => {
         expect(buttongroup instanceof IgxButtonGroupComponent).toBe(true);
         expect(buttongroup.disabled).toBeFalsy();
         expect(buttongroup.alignment).toBe(ButtonGroupAlignment.vertical);
-        expect(buttongroup.multiSelection).toBeTruthy();
+        expect(buttongroup.selectionMode).toBe('multi');
         expect(buttongroup.itemContentCssClass).toEqual('customContentStyle');
         expect(buttongroup.selectedIndexes.length).toEqual(0);
         expect(buttongroup.selectedButtons.length).toEqual(0);
@@ -197,12 +198,32 @@ describe('IgxButtonGroup', () => {
         expect(buttongroup.buttons.indexOf(buttongroup.selectedButtons[0])).toBe(2);
     });
 
+    it('Button Group single required selection', () => {
+        const fixture = TestBed.createComponent(InitButtonGroupComponent);
+        fixture.detectChanges();
+        
+        const buttongroup = fixture.componentInstance.buttonGroup;
+        buttongroup.selectionMode = 'singleRequired';
+        spyOn(buttongroup.deselected, 'emit');
+        
+        buttongroup.selectButton(0);
+        expect(buttongroup.selectedButtons.length).toBe(1);
+        expect(buttongroup.buttons.indexOf(buttongroup.selectedButtons[0])).toBe(0);
+        
+        const button = fixture.debugElement.nativeElement.querySelector('button');
+        button.click();
+
+        expect(buttongroup.selectedButtons.length).toBe(1);
+        expect(buttongroup.buttons.indexOf(buttongroup.selectedButtons[0])).toBe(0);
+        expect(buttongroup.deselected.emit).not.toHaveBeenCalled();
+    });
+
    it('Button Group multiple selection', () => {
         const fixture = TestBed.createComponent(InitButtonGroupWithValuesComponent);
         fixture.detectChanges();
 
         const buttongroup = fixture.componentInstance.buttonGroup;
-        expect(buttongroup.multiSelection).toBeTruthy();
+        expect(buttongroup.selectionMode).toBe('multi');
         buttongroup.selectButton(1);
         expect(buttongroup.selectedButtons.length).toBe(1);
         buttongroup.selectButton(2);
@@ -221,7 +242,7 @@ describe('IgxButtonGroup', () => {
         fixture.detectChanges();
 
         const buttongroup = fixture.componentInstance.buttonGroup;
-        expect(buttongroup.multiSelection).toBeTruthy();
+        expect(buttongroup.selectionMode).toBe('multi');
 
         UIInteractions.simulateClickEvent(buttongroup.buttons[0].nativeElement);
         expect(buttongroup.selectedButtons.length).toBe(1);
@@ -243,7 +264,7 @@ describe('IgxButtonGroup', () => {
         const buttongroup = fixture.componentInstance.buttonGroup;
         expect(buttongroup.buttons.length).toBe(4);
 
-        expect(buttongroup.multiSelection).toBeTruthy();
+        expect(buttongroup.selectionMode).toBe('multi');
         buttongroup.selectButton(1);
         expect(buttongroup.selectedButtons.length).toBe(1);
         buttongroup.selectButton(2);
@@ -261,12 +282,12 @@ describe('IgxButtonGroup', () => {
 
     it('Button Group - templated buttons with single selection', () => {
         const fixture = TestBed.createComponent(TemplatedButtonGroupComponent);
-        fixture.componentInstance.multiselection = false;
         fixture.detectChanges();
 
         const buttongroup = fixture.componentInstance.buttonGroup;
+        buttongroup.selectionMode = 'single';
         expect(buttongroup.buttons.length).toBe(4);
-        expect(buttongroup.multiSelection).toBeFalsy();
+        expect(buttongroup.selectionMode).toBe('single');
 
         buttongroup.selectButton(1);
         expect(buttongroup.selectedButtons.length).toBe(1);
@@ -319,7 +340,7 @@ describe('IgxButtonGroup', () => {
         fixture.detectChanges();
 
         expect(buttongroup.displayDensity).toBe('compact', 'DisplayDensity not set!');
-        expect(buttongroup.buttons[1].element.nativeElement.classList.contains('igx-button--compact')).toBe(true, 'Missing density class!');
+        expect(getComponentSize(buttongroup.buttons[1].nativeElement)).toBe('1', 'Missing density class!');
     });
 
     it('Button Group - DisplayDensity property is applied to templated buttons', () => {
@@ -333,8 +354,8 @@ describe('IgxButtonGroup', () => {
         const groupChildren = buttongroup.buttons;
         // The density class should be applied only to buttons with no DisplayDensity set
         expect(groupChildren[0].displayDensity).toBe('compact');
-        expect(groupChildren[0].element.nativeElement.classList.contains('igx-button--compact')).toBe(true, 'Missing density class!');
-        expect(groupChildren[1].element.nativeElement.classList.contains('igx-button--cosy')).toBe(true, 'Missing density class!');
+        expect(getComponentSize(groupChildren[0].element.nativeElement)).toBe('1', 'Missing density class!');
+        expect(getComponentSize(groupChildren[1].element.nativeElement)).toBe('2', 'Missing density class!');
     });
 
     it('Button Group - should support tab navigation', () => {
@@ -391,7 +412,7 @@ class InitButtonGroupComponent implements OnInit {
 
 @Component({
     template: `
-    <igx-buttongroup [multiSelection]="true" itemContentCssClass="customContentStyle"
+    <igx-buttongroup [selectionMode]="'multi'" itemContentCssClass="customContentStyle"
         [values]="cities" [alignment]="alignment">
     </igx-buttongroup>
     `,
@@ -438,7 +459,7 @@ class InitButtonGroupWithValuesComponent implements OnInit {
 
 @Component({
     template: `
-    <igx-buttongroup [multiSelection]="multiselection" [alignment]="alignment">
+    <igx-buttongroup [selectionMode]="'multi'" [alignment]="alignment">
         <button igxButton>Sofia</button>
         <button igxButton>London</button>
         <button igxButton>New York</button>
@@ -452,12 +473,11 @@ class TemplatedButtonGroupComponent {
     @ViewChild(IgxButtonGroupComponent, { static: true }) public buttonGroup: IgxButtonGroupComponent;
 
     public alignment = ButtonGroupAlignment.vertical;
-    public multiselection = true;
 }
 
 @Component({
     template: `
-    <igx-buttongroup [multiSelection]="multiselection" displayDensity="cosy">
+    <igx-buttongroup [selectionMode]="'multi'" displayDensity="cosy">
         <button igxButton displayDensity="compact">Sofia</button>
         <button igxButton>London</button>
     </igx-buttongroup>