+++ CHANGELOG.md
@@ -39,6 +39,8 @@ All notable changes for each version of this project will be documented in this 
     When the combo components are configured with it, filtering for **"resume"** will match both **"resume"** and **"résumé"**.
 - `IgxButtonGroup`:
     - Added `owner` to the `IButtonGroupEventArgs` to identify the emitting button group instance.
+    - **Breaking Change** Added the `selectionMode` property that sets the selection mode of the buttons in the `IgxButtonGroup`. Selection modes are `single`, `singleRequired` and `multi` as default is `single`.
+    - **Breaking Change** Deprecated the `multiSelection` property and all references have been migrated to `selectionMode="multi"`.
 - `Themes`:
     - Include a standalone theme for the `igxLabel` directive to allow usage with components outside the Input Group.
 
@@ -50,6 +52,31 @@ All notable changes for each version of this project will be documented in this 
     - **Behavioral Change** When there are already grouped columns, the group drop area now shows after dragging of a column starts and not when only click actions are performed.
 - `IgxCombo`, `IgxSimpleCombo`:
     - **Breaking Change**  The `selection` property returns an array of the selected items even when a value key is provided and the `value` property returns an array of value keys instead of display keys. Automatic migrations are available and will be applied on `ng update`.
+- **Deprecation** - The `DisplayDensityToken` injection token has been deprecated. Consequently, the `displayDensity` input for all components injecting the `DisplayDensityToken` is also deprecated. Changing the size of the component(s) should be done via a new custom CSS property - `--ig-size`;
+    - Code example:
+    ```typescript
+    // *.component.ts
+    // remove the provider declaration for `DisplayDensityToken`
+    providers: [{ provide: DisplayDensityToken, useValue: { displayDensity: DisplayDensity.compact } }],
+    ```
+
+    ```html
+    <!-- Remove `[displayDensity]="'compact'"` -->
+    <igx-grid [displayDensity]="'compact'">...</igx-grid>
+    ```
+
+    ```css
+    /* 
+    Add --ig-size to a component or global file.
+    Available values are:
+    compact: --ig-size-small
+    cosy: --ig-size-medium
+    comfortable: --ig-size-large
+    */
+    igx-grid {
+        --ig-size: var(--ig-size-small);
+    }
+    ```
 
 ### New Features
 - `IgxCombo`, `IgxSimpleCombo`